change invariant contract to use sails
