+++ io/Cargo.toml
@@ -8,3 +8,4 @@ gmeta.workspace = true
 gstd.workspace = true
 contracts.workspace = true
 math.workspace = true
+sails-rtl.workspace = true
