+++ src/contracts/collections/pools.rs
@@ -1,5 +1,5 @@
 use crate::{Pool, PoolKey, InvariantError};
-use gstd::collections::HashMap;
+use sails_rtl::collections::HashMap;
 
 #[derive(Debug, Default)]
 pub struct Pools {
@@ -48,7 +48,7 @@ mod tests {
   use crate::FeeTier;
   use math::percentage::Percentage;
   use decimal::*;
-  use gstd::ActorId;
+  use sails_rtl::ActorId;
 
   #[test]
   fn test_add() {
