+++ src/contracts/storage/mod.rs
@@ -3,9 +3,13 @@ pub mod pool_key;
 pub mod pool;
 pub mod tick;
 pub mod position;
+pub mod utils;
+pub mod awaiting_transfers;
 
 pub use fee_tier::*;
 pub use pool_key::*;
 pub use pool::*;
 pub use tick::*;
 pub use position::*;
+pub use awaiting_transfers::*;
+
