+++ src/contracts/storage/utils.rs
@@ -1,0 +1,55 @@
+#[macro_export]
+macro_rules! declare_storage {
+    (name: $name: ident, ty: $ty: ty $(,)?) => {
+        $crate::declare_storage!(module: internal, name: $name, ty: $ty);
+    };
+
+    (module: $module: ident, name: $name: ident, ty: $ty: ty $(,)?) => {
+        pub struct $name(());
+
+        pub use $module::*;
+
+        mod $module {
+            use super::*;
+
+            static mut INSTANCE: Option<$ty> = None;
+
+            impl $name {
+                pub fn is_set() -> bool {
+                    unsafe { INSTANCE.is_some() }
+                }
+
+                pub fn set(value: $ty) -> Result<(), $ty> {
+                    if Self::is_set() {
+                        Err(value)
+                    } else {
+                        unsafe { INSTANCE = Some(value) }
+                        Ok(())
+                    }
+                }
+
+                pub fn as_ref() -> &'static $ty {
+                    unsafe {
+                        INSTANCE.as_ref().unwrap_or_else(|| {
+                            panic!(
+                                "Storage {} should be set before accesses",
+                                stringify!($name)
+                            )
+                        })
+                    }
+                }
+
+                pub fn as_mut() -> &'static mut $ty {
+                    unsafe {
+                        INSTANCE.as_mut().unwrap_or_else(|| {
+                            panic!(
+                                "Storage {} should be set before accesses",
+                                stringify!($name)
+                            )
+                        })
+                    }
+                }
+            }
+        }
+    };
+}
