+++ src/wasm/invariant.idl
@@ -1,0 +1,172 @@
+type InvariantConfig = struct {
+  admin: actor_id,
+  protocol_fee: Percentage,
+};
+
+type Percentage = struct {
+  u64,
+};
+
+type FeeTier = struct {
+  fee: Percentage,
+  tick_spacing: u16,
+};
+
+type InvariantError = enum {
+  NotAdmin,
+  NotFeeReceiver,
+  PoolAlreadyExist,
+  PoolNotFound,
+  TickAlreadyExist,
+  InvalidTickIndexOrTickSpacing,
+  PositionNotFound,
+  TickNotFound,
+  FeeTierNotFound,
+  PoolKeyNotFound,
+  AmountIsZero,
+  WrongLimit,
+  PriceLimitReached,
+  NoGainSwap,
+  InvalidTickSpacing,
+  FeeTierAlreadyExist,
+  PoolKeyAlreadyExist,
+  UnauthorizedFeeReceiver,
+  ZeroLiquidity,
+  RecoverableTransferError,
+  UnrecoverableTransferError,
+  TransferError,
+  TokensAreSame,
+  AmountUnderMinimumAmountOut,
+  InvalidFee,
+  NotEmptyTickDeinitialization,
+  InvalidInitTick,
+  InvalidInitSqrtPrice,
+  NotEnoughGasToExecute,
+  TickLimitReached,
+  InvalidTickIndex,
+  NoBalanceForTheToken,
+  FailedToChangeTokenBalance,
+  ReplyHandlingFailed,
+};
+
+type PoolKey = struct {
+  token_x: actor_id,
+  token_y: actor_id,
+  fee_tier: FeeTier,
+};
+
+type TokenAmount = struct {
+  u128,
+};
+
+type SqrtPrice = struct {
+  u128,
+};
+
+type Liquidity = struct {
+  u128,
+};
+
+type Position = struct {
+  pool_key: PoolKey,
+  liquidity: Liquidity,
+  lower_tick_index: i32,
+  upper_tick_index: i32,
+  fee_growth_inside_x: FeeGrowth,
+  fee_growth_inside_y: FeeGrowth,
+  last_block_number: u64,
+  tokens_owed_x: TokenAmount,
+  tokens_owed_y: TokenAmount,
+};
+
+type FeeGrowth = struct {
+  u128,
+};
+
+type SwapHop = struct {
+  pool_key: PoolKey,
+  x_to_y: bool,
+};
+
+type CalculateSwapResult = struct {
+  amount_in: TokenAmount,
+  amount_out: TokenAmount,
+  start_sqrt_price: SqrtPrice,
+  target_sqrt_price: SqrtPrice,
+  fee: TokenAmount,
+  pool: Pool,
+  ticks: vec Tick,
+};
+
+type Pool = struct {
+  liquidity: Liquidity,
+  sqrt_price: SqrtPrice,
+  current_tick_index: i32,
+  fee_growth_global_x: FeeGrowth,
+  fee_growth_global_y: FeeGrowth,
+  fee_protocol_token_x: TokenAmount,
+  fee_protocol_token_y: TokenAmount,
+  start_timestamp: u64,
+  last_timestamp: u64,
+  fee_receiver: actor_id,
+};
+
+type Tick = struct {
+  index: i32,
+  sign: bool,
+  liquidity_change: Liquidity,
+  liquidity_gross: Liquidity,
+  sqrt_price: SqrtPrice,
+  fee_growth_outside_x: FeeGrowth,
+  fee_growth_outside_y: FeeGrowth,
+  seconds_outside: u64,
+};
+
+type QuoteResult = struct {
+  amount_in: TokenAmount,
+  amount_out: TokenAmount,
+  target_sqrt_price: SqrtPrice,
+  ticks: vec Tick,
+};
+
+constructor {
+  New : (config: InvariantConfig);
+};
+
+service Service {
+  AddFeeTier : (fee_tier: FeeTier) -> result (FeeTier, InvariantError);
+  ChangeFeeReceiver : (pool_key: PoolKey, fee_receiver: actor_id) -> result (null, InvariantError);
+  ChangeProtocolFee : (protocol_fee: Percentage) -> result (Percentage, InvariantError);
+  ClaimFee : (index: u32) -> result (struct { TokenAmount, TokenAmount }, InvariantError);
+  CreatePool : (token_x: actor_id, token_y: actor_id, fee_tier: FeeTier, init_sqrt_price: SqrtPrice, init_tick: i32) -> result (null, InvariantError);
+  CreatePosition : (pool_key: PoolKey, lower_tick: i32, upper_tick: i32, liquidity_delta: Liquidity, slippage_limit_lower: SqrtPrice, slippage_limit_upper: SqrtPrice) -> result (Position, InvariantError);
+  DepositSingleToken : (token: actor_id, caller: actor_id, amount: TokenAmount) -> result (TokenAmount, InvariantError);
+  DepositTokenPair : (caller: actor_id, token_x: struct { actor_id, TokenAmount }, token_y: struct { actor_id, TokenAmount }) -> result (struct { TokenAmount, TokenAmount }, InvariantError);
+  QuoteRoute : (amount_in: TokenAmount, swaps: vec SwapHop) -> result (TokenAmount, InvariantError);
+  RemoveFeeTier : (fee_tier: FeeTier) -> result (FeeTier, InvariantError);
+  RemovePosition : (index: u32) -> result (struct { TokenAmount, TokenAmount }, InvariantError);
+  Swap : (pool_key: PoolKey, x_to_y: bool, amount: TokenAmount, by_amount_in: bool, sqrt_price_limit: SqrtPrice) -> result (CalculateSwapResult, InvariantError);
+  TransferPosition : (index: u32, receiver: actor_id) -> result (null, InvariantError);
+  WithdrawProtocolFee : (pool_key: PoolKey) -> result (null, InvariantError);
+  WithdrawSingleToken : (token: actor_id, caller: actor_id, amount: opt TokenAmount) -> result (TokenAmount, InvariantError);
+  WithdrawTokenPair : (caller: actor_id, token_x: struct { actor_id, opt TokenAmount }, token_y: struct { actor_id, opt TokenAmount }) -> result (struct { TokenAmount, TokenAmount }, InvariantError);
+  query FeeTierExists : (fee_tier: FeeTier) -> bool;
+  query GetAllPositions : (owner_id: actor_id) -> vec Position;
+  query GetFeeTiers : () -> vec FeeTier;
+  query GetPool : (token_x: actor_id, token_y: actor_id, fee_tier: FeeTier) -> result (Pool, InvariantError);
+  query GetPools : (size: u8, offset: u16) -> result (vec PoolKey, InvariantError);
+  query GetPosition : (owner_id: actor_id, index: u32) -> result (Position, InvariantError);
+  query GetProtocolFee : () -> Percentage;
+  query GetTick : (key: PoolKey, index: i32) -> result (Tick, InvariantError);
+  query GetUserBalances : (user: actor_id) -> vec struct { actor_id, TokenAmount };
+  query IsTickInitialized : (key: PoolKey, index: i32) -> bool;
+  query Quote : (pool_key: PoolKey, x_to_y: bool, amount: TokenAmount, by_amount_in: bool, sqrt_price_limit: SqrtPrice) -> result (QuoteResult, InvariantError);
+
+  events {
+    PositionCreatedEvent: struct { block_timestamp: u64, address: actor_id, pool_key: PoolKey, liquidity_delta: Liquidity, lower_tick: i32, upper_tick: i32, current_sqrt_price: SqrtPrice };
+    PositionRemovedEvent: struct { block_timestamp: u64, caller: actor_id, pool_key: PoolKey, liquidity: Liquidity, lower_tick_index: i32, upper_tick_index: i32, sqrt_price: SqrtPrice };
+    CrossTickEvent: struct { timestamp: u64, address: actor_id, pool: PoolKey, indexes: vec i32 };
+    SwapEvent: struct { timestamp: u64, address: actor_id, pool: PoolKey, amount_in: TokenAmount, amount_out: TokenAmount, fee: TokenAmount, start_sqrt_price: SqrtPrice, target_sqrt_price: SqrtPrice, x_to_y: bool };
+  }
+};
+
