+++ src/wasm/src/lib.rs
@@ -1,0 +1,5 @@
+#![no_std]
+
+#[allow(unused_imports)]
+#[cfg(target_arch = "wasm32")]
+pub use invariant_app::wasm::*;
